Peter [mailto:petereverett_users.sf.net]
added to DisplayDevices.h
case 145
     Hserprint "Maxim/Dallas device"
case 147
     Hserprint "Maxim/Dallas device"
 case 149
     Hserprint "Maxim/Dallas device"
 case 151
     Hserprint "Maxim/Dallas device"
 case 153
     Hserprint "Maxim/Dallas device"
 case 155
     Hserprint "Maxim/Dallas device"
 case 157
     Hserprint "Maxim/Dallas device"
 case 159
     Hserprint "Maxim/Dallas device"
case 223
     Hserprint "MCP7940x SRAM RTC device"
 case 237
     Hserprint "Bosch Sensortec device" 
 case 239
     Hserprint "Bosch Sensortec device" 
The address of the DS1261 can range from 145 to 159, however a few other Maxim devices seem to use the same range of addresses (MAX31875, MAX44009). The DS1307 is an exception.

The MCP7940x EEPROM RTCs use address 175, which identified as "EEPROM Device" which is OK.

The Bosch devices are the BMP180/280 and the BME680.